chore: rename to 'edusharing-mock'
