+++ src/edusharing-mock/mocked-embed-json/image.ts
